src/main/webapp/app/entities/therapy/delete/therapy-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { TherapyService } from '../service/therapy.service';

import { TherapyDeleteDialogComponent } from './therapy-delete-dialog.component';

describe('Therapy Management Delete Component', () => {
  let comp: TherapyDeleteDialogComponent;
  let fixture: ComponentFixture<TherapyDeleteDialogComponent>;
  let service: TherapyService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, TherapyDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(TherapyDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(TherapyDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(TherapyService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
